wip fixed all but one test
