+++ VideoWeb/VideoWeb/ClientApp/src/app/vh-officer/vho-herings-filter/vho-hearings-filter.component.spec.ts
@@ -14,13 +14,16 @@ describe('VhoHearingsFilterComponent', () => {
         'getFilter',
         'countOptions'
     ]);
-    hearingsFilterOptionsServiceSpy.getFilter.and.returnValue(hearingFilter);
+    hearingsFilterOptionsServiceSpy.getFilter.and.returnValue(Promise.resolve(hearingFilter));
     hearingsFilterOptionsServiceSpy.countOptions.and.returnValue(3);
 
     const component = new VhoHearingsFilterComponent(hearingsFilterOptionsServiceSpy);
 
+    beforeEach(async () => {
+        await component.ngOnInit();
+    });
+
     it('should get selected filter options for All and apply filter is disabled', () => {
-        component.ngOnInit();
         expect(component.hearingsFilter).toBeTruthy();
         expect(component.statusAllChecked).toBeTruthy();
         expect(component.alertsAllChecked).toBeTruthy();
@@ -28,42 +31,36 @@ describe('VhoHearingsFilterComponent', () => {
         expect(component.disableFilterApply).toBeTruthy();
     });
     it('should select filter status option and enable apply filter', () => {
-        component.ngOnInit();
         component.statusOptionSelected(0);
         expect(component.statusAllChecked).toBeFalsy();
         expect(component.hearingsFilter.statuses[0].Selected).toBe(true);
         expect(component.disableFilterApply).toBe(false);
     });
     it('should select filter location option and enable apply filter', () => {
-        component.ngOnInit();
         component.locationOptionSelected(0);
         expect(component.locationAllChecked).toBe(false);
         expect(component.hearingsFilter.locations[0].Selected).toBe(true);
         expect(component.disableFilterApply).toBe(false);
     });
     it('should select filter alerts option and enable apply filter', () => {
-        component.ngOnInit();
         component.alertOptionSelected(0);
         expect(component.alertsAllChecked).toBe(false);
         expect(component.hearingsFilter.alerts[0].Selected).toBe(true);
         expect(component.disableFilterApply).toBe(false);
     });
     it('should select all status options', () => {
-        component.ngOnInit();
         expect(component.statusAllChecked).toBeFalsy();
         component.statusAllSelected();
         expect(component.hearingsFilter.statuses[0].Selected).toBe(false);
         expect(component.statusAllChecked).toBe(true);
     });
     it('should select all location options', () => {
-        component.ngOnInit();
         expect(component.locationAllChecked).toBe(false);
         component.locationAllSelected();
         expect(component.hearingsFilter.locations[0].Selected).toBe(false);
         expect(component.locationAllChecked).toBe(true);
     });
     it('should select all alerts options', () => {
-        component.ngOnInit();
         expect(component.alertsAllChecked).toBe(false);
         component.alertAllSelected();
         expect(component.hearingsFilter.alerts[0].Selected).toBe(false);
