+++ VideoWeb/VideoWeb/ClientApp/src/app/vh-officer/vho-herings-filter/vho-hearings-filter.component.ts
@@ -21,7 +21,7 @@ export class VhoHearingsFilterComponent implements OnInit {
 
     constructor(private hearingsFilterOptionsService: HearingsFilterOptionsService) {}
 
-    ngOnInit() {
+    async ngOnInit() {
         this.hearingsFilterOptionsService.getFilter().then(filter => {
             this.hearingsFilter = filter;
             this.countOptions(false);
